Merge

+++ rust-sdk/core/src/types/fees.rs
@@ -11,6 +11,8 @@ use tsify::Tsify;
 #[cfg_attr(feature = "wasm", serde(rename_all = "camelCase"))]
 #[cfg_attr(feature = "wasm", tsify(into_wasm_abi))]
 pub struct CollectFeesQuote {
-    pub fee_owed_a: u128,
-    pub fee_owed_b: u128,
+    #[cfg_attr(feature = "wasm", tsify(type = "bigint"))]
+    pub fee_owed_a: u64,
+    #[cfg_attr(feature = "wasm", tsify(type = "bigint"))]
+    pub fee_owed_b: u64,
 }
